Adding a WPF SheetPlacementEditor. Net 5 isn't happy referencing Net 4.8 System.Text.Json; link in to Net 5. Running, loading and rendering; but it's the simplification not original points that's saved with the SheetPlacement; resource from the original file where possible commented. Wpf highlighting current selected part, zoom at crosshairs. Add MinkowskiSum decouple

+++ DeepNestLib.CiTests/DefaultSvgNestConfig.cs
@@ -1,5 +1,7 @@
 ﻿namespace DeepNestLib.CiTests
 {
+  using DeepNestLib.NestProject;
+
   public class DefaultSvgNestConfig : ISvgNestConfig
   {
     public double ClipperScale { get; set; } = 10000000;
@@ -46,9 +48,9 @@
 
     public bool UseParallel { get; set; } = false;
 
-    public float Tolerance { get; set; } = 2f;
+    public double Tolerance { get; set; } = 2;
 
-    public float ToleranceSvg { get; set; } = 0.005f;
+    public double ToleranceSvg { get; set; } = 0.005;
 
     public AnglesEnum StrictAngles { get; set; } = AnglesEnum.None;
