+++ DeepNestLib/DbCacheKey.cs
@@ -12,7 +12,7 @@
 
   public class DbCacheKey
   {
-    public DbCacheKey(int? a, int? b, float aRotation, float bRotation, IEnumerable<INfp> nfps)
+    public DbCacheKey(int? a, int? b, double aRotation, double bRotation, IEnumerable<INfp> nfps)
     {
       A = a;
       B = b;
@@ -21,7 +21,7 @@
       nfp = nfps.ToArray();
     }
 
-    public DbCacheKey(int? a, int? b, float aRotation, float bRotation)
+    public DbCacheKey(int? a, int? b, double aRotation, double bRotation)
     // : this(a, b, aRotation, bRotation, null)
     {
       A = a;
@@ -34,9 +34,9 @@
 
     public int? B { get; }
 
-    public float ARotation { get; }
+    public double ARotation { get; }
 
-    public float BRotation { get; }
+    public double BRotation { get; }
 
     public INfp[] nfp { get; }
