+++ DeepNestLib/NfpPairDictionary.cs
@@ -9,12 +9,12 @@
     {
     }
 
-    public bool TryGetValue(SvgPoint[] a, SvgPoint[] b, float aRotation, float bRotation, int aSource, int bSource, MinkowskiSumPick minkowskiSumPick, out NFP value)
+    public bool TryGetValue(SvgPoint[] a, SvgPoint[] b, double aRotation, double bRotation, int aSource, int bSource, MinkowskiSumPick minkowskiSumPick, out NFP value)
     {
       return this.TryGetValue(new NfpPairEqualityComparerKey(a, b, aRotation, bRotation, aSource, bSource, minkowskiSumPick), out value);
     }
 
-    public void Add(SvgPoint[] a, SvgPoint[] b, float aRotation, float bRotation, int aSource, int bSource, MinkowskiSumPick minkowskiSumPick, NFP value)
+    public void Add(SvgPoint[] a, SvgPoint[] b, double aRotation, double bRotation, int aSource, int bSource, MinkowskiSumPick minkowskiSumPick, NFP value)
     {
       this.Add(new NfpPairEqualityComparerKey(a, b, aRotation, bRotation, aSource, bSource, minkowskiSumPick), value);
     }
